refactor: Move node-label to its own file (reduce code complexity)
